+++ server/src/main/java/org/opensearch/indices/replication/OngoingSegmentReplications.java
@@ -41,6 +41,8 @@ import java.util.stream.Collectors;
  * @opensearch.internal
  */
 class OngoingSegmentReplications {
+
+    private static final Logger logger = LogManager.getLogger(OngoingSegmentReplications.class);
     private final RecoverySettings recoverySettings;
     private final IndicesService indicesService;
     private final Map<ReplicationCheckpoint, CopyState> copyStateMap;
@@ -95,8 +97,6 @@ class OngoingSegmentReplications {
         }
     }
 
-    private static final Logger logger = LogManager.getLogger(OngoingSegmentReplications.class);
-
     /**
      * Start sending files to the replica.
      *
@@ -121,6 +121,10 @@ class OngoingSegmentReplications {
                 }
             });
             if (request.getFilesToFetch().isEmpty()) {
+                // before completion, alert the primary of the replica's state.
+                handler.getCopyState()
+                    .getShard()
+                    .updateVisibleCheckpointForShard(request.getTargetAllocationId(), handler.getCopyState().getCheckpoint());
                 wrappedListener.onResponse(new GetSegmentFilesResponse(Collections.emptyList()));
             } else {
                 handler.sendFiles(request, wrappedListener);
@@ -265,6 +269,9 @@ class OngoingSegmentReplications {
             .filter(predicate)
             .map(SegmentReplicationSourceHandler::getAllocationId)
             .collect(Collectors.toList());
+        if (allocationIds.size() == 0) {
+            return;
+        }
         logger.warn(() -> new ParameterizedMessage("Cancelling replications for allocationIds {}", allocationIds));
         for (String allocationId : allocationIds) {
             cancel(allocationId, reason);