+++ server/src/test/java/org/opensearch/cluster/routing/allocation/AllocationConstraintsTests.java
@@ -18,9 +18,10 @@ import org.opensearch.common.settings.Settings;
 import static org.mockito.Mockito.anyString;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
-import static org.opensearch.cluster.routing.allocation.AllocationConstraints.INDEX_SHARD_PER_NODE_BREACH_CONSTRAINT_ID;
-import static org.opensearch.cluster.routing.allocation.Constraint.CONSTRAINT_WEIGHT;
-import static org.opensearch.cluster.routing.allocation.RebalanceConstraints.PREFER_PRIMARY_SHARD_BALANCE_NODE_BREACH_ID;
+import static org.opensearch.cluster.routing.allocation.ConstraintTypes.CLUSTER_PRIMARY_SHARD_BALANCE_CONSTRAINT_ID;
+import static org.opensearch.cluster.routing.allocation.ConstraintTypes.CONSTRAINT_WEIGHT;
+import static org.opensearch.cluster.routing.allocation.ConstraintTypes.INDEX_SHARD_PER_NODE_BREACH_CONSTRAINT_ID;
+import static org.opensearch.cluster.routing.allocation.ConstraintTypes.INDEX_PRIMARY_SHARD_BALANCE_CONSTRAINT_ID;
 
 public class AllocationConstraintsTests extends OpenSearchAllocationTestCase {
 
@@ -36,7 +37,7 @@ public class AllocationConstraintsTests extends OpenSearchAllocationTestCase {
         settings.put(BalancedShardsAllocator.INDEX_BALANCE_FACTOR_SETTING.getKey(), indexBalanceFactor);
         settings.put(BalancedShardsAllocator.SHARD_BALANCE_FACTOR_SETTING.getKey(), shardBalance);
         settings.put(BalancedShardsAllocator.THRESHOLD_SETTING.getKey(), threshold);
-        settings.put(BalancedShardsAllocator.PREFER_PER_INDEX_PRIMARY_SHARD_BALANCE.getKey(), true);
+        settings.put(BalancedShardsAllocator.PREFER_PRIMARY_SHARD_BALANCE.getKey(), true);
 
         service.applySettings(settings.build());
 
@@ -45,7 +46,7 @@ public class AllocationConstraintsTests extends OpenSearchAllocationTestCase {
         assertEquals(threshold, allocator.getThreshold(), 0.01);
         assertEquals(true, allocator.getPreferPrimaryBalance());
 
-        settings.put(BalancedShardsAllocator.PREFER_PER_INDEX_PRIMARY_SHARD_BALANCE.getKey(), false);
+        settings.put(BalancedShardsAllocator.PREFER_PRIMARY_SHARD_BALANCE.getKey(), false);
         service.applySettings(settings.build());
         assertEquals(false, allocator.getPreferPrimaryBalance());
     }
@@ -76,54 +77,133 @@ public class AllocationConstraintsTests extends OpenSearchAllocationTestCase {
      * Test constraint evaluation logic when with different values of ConstraintMode
      * for IndexShardPerNode constraint satisfied and breached.
      */
-    public void testIndexPrimaryShardsPerNodeConstraint() {
+    public void testPerIndexPrimaryShardsConstraint() {
         ShardsBalancer balancer = mock(LocalShardsBalancer.class);
         BalancedShardsAllocator.ModelNode node = mock(BalancedShardsAllocator.ModelNode.class);
         AllocationConstraints constraints = new AllocationConstraints();
-        constraints.updateAllocationConstraint(PREFER_PRIMARY_SHARD_BALANCE_NODE_BREACH_ID, true);
+        constraints.updateAllocationConstraint(INDEX_PRIMARY_SHARD_BALANCE_CONSTRAINT_ID, true);
 
+        final String indexName = "test-index";
+        int perIndexPrimaryShardCount = 1;
+        float avgPerIndexPrimaryShardsPerNode = 2f;
+
+        when(balancer.avgPrimaryShardsPerNode(anyString())).thenReturn(avgPerIndexPrimaryShardsPerNode);
+        when(node.numPrimaryShards(anyString())).thenReturn(perIndexPrimaryShardCount);
+        when(node.getNodeId()).thenReturn("test-node");
+
+        assertEquals(0, constraints.weight(balancer, node, indexName));
+
+        perIndexPrimaryShardCount = 3;
+        when(node.numPrimaryShards(anyString())).thenReturn(perIndexPrimaryShardCount);
+        assertEquals(CONSTRAINT_WEIGHT, constraints.weight(balancer, node, indexName));
+
+        constraints.updateAllocationConstraint(INDEX_PRIMARY_SHARD_BALANCE_CONSTRAINT_ID, false);
+        assertEquals(0, constraints.weight(balancer, node, indexName));
+    }
+
+    /**
+     * Test constraint evaluation logic when per index primary shard count constraint is breached.
+     */
+    public void testGlobalPrimaryShardsConstraint() {
+        ShardsBalancer balancer = mock(LocalShardsBalancer.class);
+        BalancedShardsAllocator.ModelNode node = mock(BalancedShardsAllocator.ModelNode.class);
+        AllocationConstraints constraints = new AllocationConstraints();
+        constraints.updateAllocationConstraint(CLUSTER_PRIMARY_SHARD_BALANCE_CONSTRAINT_ID, true);
+
+        final String indexName = "test-index";
         int primaryShardCount = 1;
         float avgPrimaryShardsPerNode = 2f;
 
-        when(balancer.avgPrimaryShardsPerNode(anyString())).thenReturn(avgPrimaryShardsPerNode);
-        when(node.numPrimaryShards(anyString())).thenReturn(primaryShardCount);
+        when(balancer.avgPrimaryShardsPerNode()).thenReturn(avgPrimaryShardsPerNode);
+        when(node.numPrimaryShards()).thenReturn(primaryShardCount);
         when(node.getNodeId()).thenReturn("test-node");
 
-        assertEquals(0, constraints.weight(balancer, node, "index"));
+        assertEquals(0, constraints.weight(balancer, node, indexName));
 
         primaryShardCount = 3;
-        when(node.numPrimaryShards(anyString())).thenReturn(primaryShardCount);
-        assertEquals(CONSTRAINT_WEIGHT, constraints.weight(balancer, node, "index"));
+        when(node.numPrimaryShards()).thenReturn(primaryShardCount);
+        assertEquals(CONSTRAINT_WEIGHT, constraints.weight(balancer, node, indexName));
 
-        constraints.updateAllocationConstraint(PREFER_PRIMARY_SHARD_BALANCE_NODE_BREACH_ID, false);
-        assertEquals(0, constraints.weight(balancer, node, "index"));
+        constraints.updateAllocationConstraint(CLUSTER_PRIMARY_SHARD_BALANCE_CONSTRAINT_ID, false);
+        assertEquals(0, constraints.weight(balancer, node, indexName));
+    }
+
+    /**
+     * Test constraint evaluation logic when both per index and global primary shard count constraint is breached.
+     */
+    public void testPrimaryShardsConstraints() {
+        ShardsBalancer balancer = mock(LocalShardsBalancer.class);
+        BalancedShardsAllocator.ModelNode node = mock(BalancedShardsAllocator.ModelNode.class);
+        AllocationConstraints constraints = new AllocationConstraints();
+        constraints.updateAllocationConstraint(CLUSTER_PRIMARY_SHARD_BALANCE_CONSTRAINT_ID, true);
+        constraints.updateAllocationConstraint(INDEX_PRIMARY_SHARD_BALANCE_CONSTRAINT_ID, true);
+
+        final String indexName = "test-index";
+        int perIndexPrimaryShardCount = 1;
+        float avgPerIndexPrimaryShardCount = 2;
+        int primaryShardCount = 2;
+        float avgPrimaryShardsPerNode = 4;
+
+        when(balancer.avgPrimaryShardsPerNode(indexName)).thenReturn(avgPerIndexPrimaryShardCount);
+        when(node.numPrimaryShards(indexName)).thenReturn(perIndexPrimaryShardCount);
+        when(balancer.avgPrimaryShardsPerNode()).thenReturn(avgPrimaryShardsPerNode);
+        when(node.numPrimaryShards()).thenReturn(primaryShardCount);
+        when(node.getNodeId()).thenReturn("test-node");
+
+        assertEquals(0, constraints.weight(balancer, node, indexName));
+
+        // breaching global primary shard count but not per index primary shard count
+        primaryShardCount = 5;
+        when(node.numPrimaryShards()).thenReturn(primaryShardCount);
+        assertEquals(CONSTRAINT_WEIGHT, constraints.weight(balancer, node, indexName));
+
+        // when per index primary shard count constraint is also breached
+        perIndexPrimaryShardCount = 3;
+        when(node.numPrimaryShards(indexName)).thenReturn(perIndexPrimaryShardCount);
+        assertEquals(2 * CONSTRAINT_WEIGHT, constraints.weight(balancer, node, indexName));
+
+        // disable both constraints
+        constraints.updateAllocationConstraint(INDEX_PRIMARY_SHARD_BALANCE_CONSTRAINT_ID, false);
+        constraints.updateAllocationConstraint(CLUSTER_PRIMARY_SHARD_BALANCE_CONSTRAINT_ID, false);
+        assertEquals(0, constraints.weight(balancer, node, indexName));
     }
 
     /**
      * Test constraint evaluation logic when with different values of ConstraintMode
      * for IndexShardPerNode constraint satisfied and breached.
      */
-    public void testAllConstraint() {
+    public void testAllConstraints() {
         ShardsBalancer balancer = mock(LocalShardsBalancer.class);
         BalancedShardsAllocator.ModelNode node = mock(BalancedShardsAllocator.ModelNode.class);
         AllocationConstraints constraints = new AllocationConstraints();
         constraints.updateAllocationConstraint(INDEX_SHARD_PER_NODE_BREACH_CONSTRAINT_ID, true);
-        constraints.updateAllocationConstraint(PREFER_PRIMARY_SHARD_BALANCE_NODE_BREACH_ID, true);
+        constraints.updateAllocationConstraint(INDEX_PRIMARY_SHARD_BALANCE_CONSTRAINT_ID, true);
+        constraints.updateAllocationConstraint(CLUSTER_PRIMARY_SHARD_BALANCE_CONSTRAINT_ID, true);
 
+        final String indexName = "test-index";
         int shardCount = randomIntBetween(1, 500);
-        int primaryShardCount = randomIntBetween(1, shardCount);
-        float avgShardsPerNode = 1.0f + (random().nextFloat()) * 999.0f;
-        float avgPrimaryShardsPerNode = (random().nextFloat()) * avgShardsPerNode;
+        float avgPerIndexShardsPerNode = 1.0f + (random().nextFloat()) * 999.0f;
 
-        when(balancer.avgPrimaryShardsPerNode(anyString())).thenReturn(avgPrimaryShardsPerNode);
-        when(node.numPrimaryShards(anyString())).thenReturn(primaryShardCount);
-        when(balancer.avgShardsPerNode(anyString())).thenReturn(avgShardsPerNode);
-        when(node.numShards(anyString())).thenReturn(shardCount);
+        int perIndexPrimaryShardCount = randomIntBetween(1, shardCount);
+        float avgPerIndexPrimaryShardsPerNode = (random().nextFloat()) * avgPerIndexShardsPerNode;
+
+        float avgPrimaryShardsPerNode = 1.0f + (random().nextFloat()) * 999.0f;
+        int primaryShardsPerNode = randomIntBetween(1, shardCount);
+
+        when(balancer.avgPrimaryShardsPerNode(indexName)).thenReturn(avgPerIndexPrimaryShardsPerNode);
+        when(node.numPrimaryShards(indexName)).thenReturn(perIndexPrimaryShardCount);
+
+        when(balancer.avgPrimaryShardsPerNode()).thenReturn(avgPrimaryShardsPerNode);
+        when(node.numPrimaryShards()).thenReturn(primaryShardsPerNode);
+
+        when(balancer.avgShardsPerNode(indexName)).thenReturn(avgPerIndexShardsPerNode);
+        when(node.numShards(indexName)).thenReturn(shardCount);
         when(node.getNodeId()).thenReturn("test-node");
 
-        long expectedWeight = (shardCount >= avgShardsPerNode) ? CONSTRAINT_WEIGHT : 0;
-        expectedWeight += primaryShardCount > avgPrimaryShardsPerNode ? CONSTRAINT_WEIGHT : 0;
-        assertEquals(expectedWeight, constraints.weight(balancer, node, "index"));
+        long expectedWeight = (shardCount >= avgPerIndexShardsPerNode) ? CONSTRAINT_WEIGHT : 0;
+        expectedWeight += perIndexPrimaryShardCount > avgPerIndexPrimaryShardsPerNode ? CONSTRAINT_WEIGHT : 0;
+        expectedWeight += primaryShardsPerNode >= avgPrimaryShardsPerNode ? CONSTRAINT_WEIGHT : 0;
+        assertEquals(expectedWeight, constraints.weight(balancer, node, indexName));
     }
 
 }